Consolidated MudBaseInput<T> and MudForm
