+++ src/MudBlazor/MudBlazor.xml
@@ -81,119 +81,119 @@
             Causes this component to validate its value
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Disabled">
+        <member name="P:MudBlazor.MudBaseInput`1.Disabled">
             <summary>
             If true, the input element will be disabled.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.ReadOnly">
+        <member name="P:MudBlazor.MudBaseInput`1.ReadOnly">
             <summary>
             If true, the input will be read only.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.FullWidth">
+        <member name="P:MudBlazor.MudBaseInput`1.FullWidth">
             <summary>
             If true, the input will take up the full width of its container.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Immediate">
+        <member name="P:MudBlazor.MudBaseInput`1.Immediate">
             <summary>
             If true, the input will update the Value immediately on typing.
             If false, the Value is updated only on Enter.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.DisableUnderLine">
+        <member name="P:MudBlazor.MudBaseInput`1.DisableUnderLine">
             <summary>
             If true, the input will not have an underline.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Label">
+        <member name="P:MudBlazor.MudBaseInput`1.Label">
             <summary>
             If string has value the label text will be displayed in the input, and scaled down at the top if the input has value.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Placeholder">
+        <member name="P:MudBlazor.MudBaseInput`1.Placeholder">
             <summary>
             The short hint displayed in the input before the user enters a value.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.HelperText">
+        <member name="P:MudBlazor.MudBaseInput`1.HelperText">
             <summary>
             The HelperText will be displayed below the text field.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.AdornmentIcon">
+        <member name="P:MudBlazor.MudBaseInput`1.AdornmentIcon">
             <summary>
             Icon that will be used if Adornment is set to Start or End.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.AdornmentText">
+        <member name="P:MudBlazor.MudBaseInput`1.AdornmentText">
             <summary>
             Text that will be used if Adornment is set to Start or End, the Text overrides Icon.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Adornment">
+        <member name="P:MudBlazor.MudBaseInput`1.Adornment">
             <summary>
             Sets Start or End Adornment if not set to None.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.IconSize">
+        <member name="P:MudBlazor.MudBaseInput`1.IconSize">
             <summary>
             Sets the Icon Size.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.OnAdornmentClick">
+        <member name="P:MudBlazor.MudBaseInput`1.OnAdornmentClick">
             <summary>
             Button click event if set and Adornment used.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.InputType">
+        <member name="P:MudBlazor.MudBaseInput`1.InputType">
             <summary>
             Type of the input element. It should be a valid HTML5 input type.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Variant">
+        <member name="P:MudBlazor.MudBaseInput`1.Variant">
             <summary>
             Variant to use.
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.Margin">
+        <member name="P:MudBlazor.MudBaseInput`1.Margin">
             <summary>
              Will adjust vertical spacing. 
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.OnBlur">
+        <member name="P:MudBlazor.MudBaseInput`1.OnBlur">
             <summary>
             Fired when the Value property changes. 
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.EditContext">
+        <member name="P:MudBlazor.MudBaseInput`1.EditContext">
             <summary>
             This is the form validation context for Blazor's <EditForm></EditForm> component
             </summary>
         </member>
-        <member name="P:MudBlazor.MudBaseInputText.For">
+        <member name="P:MudBlazor.MudBaseInput`1.For">
             <summary>
             Specify an expression which returns the model's field for which validation messages should be displayed.
             Currently only string fields are supported.
             </summary>
         </member>
-        <member name="F:MudBlazor.MudBaseInputText._fieldIdentifier">
+        <member name="F:MudBlazor.MudBaseInput`1._fieldIdentifier">
             <summary>
             Points to a field of the model for which validation messages should be displayed.
             </summary>
         </member>
-        <member name="F:MudBlazor.MudBaseInputText._currentFor">
+        <member name="F:MudBlazor.MudBaseInput`1._currentFor">
             <summary>
             To find out whether or not For parameter has changed we keep a separate reference
             </summary>
         </member>
-        <member name="F:MudBlazor.MudBaseInputText._currentEditContext">
+        <member name="F:MudBlazor.MudBaseInput`1._currentEditContext">
             <summary>
             To find out whether or not EditContext parameter has changed we keep a separate reference
             </summary>
         </member>
-        <member name="M:MudBlazor.MudBaseInputText.Dispose(System.Boolean)">
+        <member name="M:MudBlazor.MudBaseInput`1.Dispose(System.Boolean)">
             <summary>
             Called to dispose this instance.
             </summary>
